Move Framework section to separate component
